+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+# [1.3.0-beta5] - 2020-01-14
 - Added `db_producer.insert` and `db_producer.process` metrics.
 
 # [1.3.0-beta4] - 2019-12-02